moving components from global components folder to wallet components
