+++ src/routes/app/wallets/_components/CoinsToApprove.tsx
